Improved demos
