+++ vaadin-maps-leaflet-flow-demo/src/main/java/software/xdev/vaadin/maps/leaflet/flow/demo/ComplexDemo.java
@@ -1,24 +1,17 @@
 package software.xdev.vaadin.maps.leaflet.flow.demo;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Random;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicLong;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.vaadin.flow.component.ClientCallable;
-import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.dialog.Dialog;
 import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
-import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.router.Route;
 
 import elemental.json.JsonObject;
@@ -52,25 +45,24 @@ import software.xdev.vaadin.maps.leaflet.map.LMapLocateOptions;
 import software.xdev.vaadin.maps.leaflet.registry.LDefaultComponentManagementRegistry;
 
 
-@Route("")
-public class LeafletView extends VerticalLayout
+@Route(ComplexDemo.NAV)
+public class ComplexDemo extends AbstractDemo
 {
-	private static final Logger LOG = LoggerFactory.getLogger(LeafletView.class);
+	public static final String NAV = "/complex";
+	
+	private static final Logger LOG = LoggerFactory.getLogger(ComplexDemo.class);
 	private static final String ID = "leaflet-demo-view";
 	
 	private final LDefaultComponentManagementRegistry reg;
 	private final LMap map;
 	private final HorizontalLayout hlButtons = new HorizontalLayout();
 	
-	public LeafletView()
+	public ComplexDemo()
 	{
-		// Add an Id to the current view so that we can later find it in JS and do a callback
 		this.setId(ID);
 		
-		// Create the registry which allows reusing components and invoking methods
 		this.reg = new LDefaultComponentManagementRegistry(this);
 		
-		// Create and add the MapContainer (which will contain the map) to the UI
 		final MapContainer mapContainer = new MapContainer(this.reg);
 		mapContainer.setSizeFull();
 		this.add(mapContainer);
@@ -129,7 +121,7 @@ public class LeafletView extends VerticalLayout
 				""")
 			.withIconSize(new LPoint(this.reg, 125, 25)));
 		
-		final LLatLng locationXDEV = new LLatLng(this.reg, 49.675806, 12.1609901);
+		final LLatLng locationXDEV = new LLatLng(this.reg, 49.6756, 12.1610);
 		final LMarker markerXDEV =
 			new LMarker(this.reg, locationXDEV, new LMarkerOptions().withIcon(iconXDEV))
 				.bindPopup("<a href='https://xdev.software' target='_blank'>XDEV Software GmbH</a>");
@@ -192,7 +184,6 @@ public class LeafletView extends VerticalLayout
 		this.addImageOverlayDemo();
 		this.addVideoOverlayDemo();
 		this.addComplexPolygonDemo();
-		this.addMemoryTestDemo();
 		
 		this.addOpenDialogOverMapDemo();
 	}
@@ -372,116 +363,6 @@ public class LeafletView extends VerticalLayout
 		));
 	}
 	
-	/**
-	 * Used for testing if memory on the client side is freed up correctly when handling a lot of components
-	 */
-	// S5413 - Yes it's used correctly
-	// S2245 - This is a reproducible demo
-	// S1215 - This is a memory test and we don't rely on random GC collects
-	@SuppressWarnings({"java:S5413", "java:S2245", "java:S1215", "java:S3776"})
-	private void addMemoryTestDemo()
-	{
-		final AtomicBoolean abort = new AtomicBoolean(false);
-		
-		this.hlButtons.add(this.createToggleButton(
-			"Start browser memory test",
-			"Stop browser memory test",
-			() -> {
-				this.map.setView(new LLatLng(this.reg, 0, 0), 12);
-				
-				abort.set(false);
-				
-				final int bulkSize = 100;
-				final int sizeWhenStartRemoving = 1_000;
-				final int endWhen = 100_000;
-				
-				final AtomicLong totallyAdded = new AtomicLong(0);
-				final Random random = new Random(1);
-				final UI ui = UI.getCurrent();
-				
-				CompletableFuture.runAsync(() -> {
-					try
-					{
-						final List<LMarker> markers = new ArrayList<>();
-						while(!abort.get() && totallyAdded.get() < endWhen)
-						{
-							LOG.info("Totally added markers: {}; Adding bulk", totallyAdded);
-							
-							for(int i = 0; i < bulkSize; i++)
-							{
-								if(markers.size() > sizeWhenStartRemoving)
-								{
-									final LMarker removed = markers.remove(0);
-									
-									ui.accessSynchronously(removed::remove);
-								}
-								final double lat = random.nextDouble(-100, 100);
-								final double lng = random.nextDouble(-100, 100);
-								final double v = random.nextDouble();
-								
-								ui.accessSynchronously(() ->
-									markers.add(
-										new LMarker(
-											this.reg,
-											new LLatLng(
-												this.reg,
-												lat,
-												lng))
-											.bindPopup("Random double: " + v)
-											.addTo(this.map)));
-							}
-							
-							totallyAdded.set(totallyAdded.get() + bulkSize);
-							
-							LOG.info("Added {}x markers; Performing GC", bulkSize);
-							System.gc();
-							LOG.info("Finished GC; Syncing with client");
-							ui.accessSynchronously(this.reg::freeUpClient);
-							
-							if(!abort.get())
-							{
-								LOG.info("Waiting a moment");
-								Thread.sleep(10);
-							}
-						}
-						
-						LOG.info("Ending; Clearing...");
-						markers.forEach(m -> ui.accessSynchronously(m::remove));
-						markers.clear();
-						
-						for(int i = 0; i < 2; i++)
-						{
-							System.gc();
-							
-							Thread.sleep(100);
-							
-							ui.accessSynchronously(this.reg::freeUpClient);
-						}
-						
-						LOG.info("Done - Browser should no longer allocate a lot of RAM now");
-						
-						ui.accessSynchronously(() -> this.reg.execJs(
-							"alert('Finished memory test!'"
-								+ "+'\\nTracked components on client (less is better): ' "
-								+ "+ document.getElementById('"
-								+ this.reg.getId().orElseThrow()
-								+ "').lComponents.size);"));
-					}
-					catch(final InterruptedException ex)
-					{
-						Thread.currentThread().interrupt();
-						LOG.warn("Got interrupted", ex);
-					}
-					catch(final Exception ex)
-					{
-						LOG.error("Unexpected problem while testing", ex);
-					}
-				});
-			},
-			() -> abort.set(true)
-		));
-	}
-	
 	private void addOpenDialogOverMapDemo()
 	{
 		this.hlButtons.add(
@@ -498,28 +379,5 @@ public class LeafletView extends VerticalLayout
 				
 				icoClose.addClickListener(iev -> dialog.close());
 			}));
-	}
-	
-	private Button createToggleButton(
-		final String showText,
-		final String hideText,
-		final Runnable onShow,
-		final Runnable onHide)
-	{
-		final AtomicBoolean shown = new AtomicBoolean(false);
-		final Button btn = new Button(
-			showText,
-			ev -> {
-				final boolean isShow = !shown.get();
-				(isShow ? onShow : onHide).run();
-				ev.getSource().setText(isShow ? hideText : showText);
-				
-				shown.set(isShow);
-				
-				ev.getSource().setEnabled(true);
-			}
-		);
-		btn.setDisableOnClick(true);
-		return btn;
 	}
 }
